Implement fixed height search/loading interface for delegate voting.
This implements a static 3 result placeholder design for the delegate
search results. This is to avoid the dialog from changing size (height)
constantly when the user is entering the search term. It's not the most
elegant solution, but miles better than the constant jumping around that
any other solution would've produced.

See #10