Fix results/suggestions label updating before content.
The label in vote delegate form was updated before the actual content
update happened, thus for 500ms the label and content were out of sync.
It was due to the search method being throttled and the form was
determining the state itself instead of relying on the parent to pass in
the correct state associated with the content it passes in.

See #10