Use explicit RawAmount type for handling of monetary values.
Fixes #5:

RawAmount and UnitAmount are backed by BigNumber, so various
overflow/impercision errors are avoided which can happen with integers
and floats. Also TypeScript will help spot incorrect usages now.

Fixes #57:

The amountToServer function (which was removed) was also the reason why
it was impossible to send amounts with decimal places. Now that all the
amounts are backed by BigNumber and it's percise conversions this is
solved.