Fix getContacts not respecting wallet accounts as primary source of truth.
The getContacts method is supposed to ignore the addressBook entry when the account
is present in the wallet, but it was not doing that. On occasion the lower-priority
address book entry also overrid the actual data from the wallet.

In addition the unit-test was kind of broken for the getContacts, so the new test
explicitly tests for the return value to match that of the expected data.