Polish the node switcher experience.
This commit adds an icon to the node switcher option in the add account
menu and only shows the link for non-official domains. It also adds back
navigation to the selection dialog and tweaks the copy a bit. The radio
button padding is reduced so that the official nodes don't take up too
much space.

Closes #163