Stop pinging device with no channels open.
During one of the Ledger code refactors the feature of pinging the
Ledger device only when the UI needs the information was lost. This
commit reimplements that feature.

This is also needed for users of Yubikeys (for which the U2F protocol
was created for), to not bombard their non-Ledger devices with U2F
requests when it's not needed. See #174.